(REF) APIv4 FieldSpec - Extract BasicSpecTrait (name, title, description)
